Refactor TextLink back to components instead of hooks
